restart!
Back to slowly working on re-writing this theme again, it may be a while before I finish the re-write but I promise to actually finish it this time and then continue to fix bugs after.

The way I'm doing things should make it a bit easier to fix bugs along with pushing updates in the future.